Fix validation command error
