+++ lib/liveValidation/liveValidator.ts
@@ -24,13 +24,13 @@ import { RuntimeException } from "../util/validationError";
 import { inversifyGetContainer, inversifyGetInstance, TYPES } from "../inversifyUtils";
 import { setDefaultOpts } from "../swagger/loader";
 import { apiValidationErrors, ApiValidationErrorCode } from "../util/errorDefinitions";
-import { kvPairsToObject } from "../util/utils";
-import { LiveValidatorLoader, LiveValidatorLoaderOption } from "./liveValidatorLoader";
 import {
+  kvPairsToObject,
   getProviderFromPathTemplate,
   getProviderFromSpecPath,
-  OperationSearcher,
-} from "./operationSearcher";
+} from "../util/utils";
+import { LiveValidatorLoader, LiveValidatorLoaderOption } from "./liveValidatorLoader";
+import { OperationSearcher } from "./operationSearcher";
 import {
   LiveRequest,
   LiveResponse,
