+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "oav",
-  "version": "2.11.2",
+  "version": "2.11.3",
   "author": {
     "name": "Microsoft Corporation",
     "email": "azsdkteam@microsoft.com",
